временно положил конфиг в прод
