+++ src/environments/environment.prod.ts
@@ -1,3 +1,13 @@
 export const environment = {
   production: true
 };
+
+export const firebaseConfig = {
+  apiKey: "AIzaSyAlAajBI87OHvVjxptf-Jyo3IH1op90XRw",
+  authDomain: "iskona-48b8c.firebaseapp.com",
+  projectId: "iskona-48b8c",
+  storageBucket: "iskona-48b8c.appspot.com",
+  messagingSenderId: "807105753873",
+  appId: "1:807105753873:web:0bd29036d5044d7c9b92b0",
+  measurementId: "G-7Y8GCGHH2K"
+};
