+++ src/environments/environment.ts
@@ -6,15 +6,7 @@ export const environment = {
   production: false
 };
 
-export const firebaseConfig = {
-  apiKey: "AIzaSyAlAajBI87OHvVjxptf-Jyo3IH1op90XRw",
-  authDomain: "iskona-48b8c.firebaseapp.com",
-  projectId: "iskona-48b8c",
-  storageBucket: "iskona-48b8c.appspot.com",
-  messagingSenderId: "807105753873",
-  appId: "1:807105753873:web:0bd29036d5044d7c9b92b0",
-  measurementId: "G-7Y8GCGHH2K"
-};
+
 /*
  * For easier debugging in development mode, you can import the following file
  * to ignore zone related error stack frames such as `zone.run`, `zoneDelegate.invokeTask`.
